+++ rich/pretty.py
@@ -18,18 +18,23 @@ if TYPE_CHECKING:  # pragma: no cover
         Console,
         ConsoleOptions,
         HighlighterType,
+        JustifyMethod,
         OverflowMethod,
         RenderResult,
     )
 
 
 def install(
-    console: "Console" = None, no_wrap: bool = False, overflow: "OverflowMethod" = None
+    console: "Console" = None,
+    overflow: "OverflowMethod" = "ignore",
+    crop: bool = False,
 ) -> None:
     """Install automatic pretty printing in the Python REPL.
 
     Args:
-        console (Console, optional): Console instance or ``None`` to use global console. Defaults to None.
+        console (Console, optional): Console instance or ``None`` to use global console. Defaults to None.        
+        overflow (Optional[OverflowMethod], optional): Overflow method. Defaults to None.
+        crop (Optional[bool], optional): Enable cropping of long lines. Defaults to False.
     """
     from rich import get_console
 
@@ -41,9 +46,8 @@ def install(
             console.print(
                 value
                 if hasattr(value, "__rich_console__") or hasattr(value, "__rich__")
-                else pretty_repr(
-                    value, max_width=console.width, no_wrap=False, overflow=overflow
-                )
+                else pretty_repr(value, max_width=console.width, overflow=overflow),
+                crop=crop,
             )
 
     sys.displayhook = display_hook
@@ -58,14 +62,14 @@ class Pretty:
         highlighter: "HighlighterType" = None,
         *,
         indent_size: int = 4,
+        justify: "JustifyMethod" = None,
         overflow: "OverflowMethod" = None,
-        no_wrap: bool = None,
     ) -> None:
         self._object = _object
         self.highlighter = highlighter or NullHighlighter()
         self.indent_size = indent_size
+        self.justify = justify
         self.overflow = overflow
-        self.no_wrap = no_wrap
 
     def __rich_console__(
         self, console: "Console", options: "ConsoleOptions"
@@ -74,8 +78,8 @@ class Pretty:
             self._object,
             max_width=options.max_width,
             indent_size=self.indent_size,
+            justify=self.justify or options.justify,
             overflow=self.overflow or options.overflow,
-            no_wrap=pick_bool(self.no_wrap, options.no_wrap, True),
         )
         yield pretty_text
 
@@ -128,6 +132,7 @@ def pretty_repr(
     max_width: Optional[int] = 80,
     indent_size: int = 4,
     highlighter: Highlighter = None,
+    justify: "JustifyMethod" = None,
     overflow: "OverflowMethod" = None,
     no_wrap: bool = True,
 ) -> Text:
@@ -244,7 +249,10 @@ def pretty_repr(
             break  # pragma: no cover
 
     text = Text(
-        "\n".join(line.text for line in lines), overflow=overflow, no_wrap=no_wrap
+        "\n".join(line.text for line in lines),
+        justify=justify,
+        overflow=overflow,
+        no_wrap=no_wrap,
     )
     text = highlighter(text)
     return text
@@ -274,6 +282,6 @@ if __name__ == "__main__":  # pragma: no cover
     console = Console()
     from rich import print
 
-    p = Pretty(data, overflow="ellipsis")
+    p = Pretty(data, overflow="ignore")
     print(Measurement.get(console, p))
-    console.print(p)
+    console.print(p, crop=False)