+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.0] - 2020-08-14
+
+- Added crop argument to Console.print
+- Added "ignore" overflow method
+- Added multiple characters per rule @hedythedev https://github.com/willmcgugan/rich/pull/207
+
 ## [5.1.2] - 2020-08-10
 
 ### Fixed